requirements.txt
alembic==1.4.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
